:recycle: Extract filters.go from pipeline_handler.go
